Fixes cheksum failure when :md5sum option is a URL.
When the :md5sum option is a URL either it is treated as an mdsum string
failing checksum test or it is not used and default mdsum url for
default mirror is used.  In either case there is a check sum failure.

This commit fixes that issue by overriding the default md5url when mdsum
is a url.

This commit also add some tests for this bug fix as well as fixes
a bug in the download URL code that throws an error when a file request
is stubbed.